sandbox area component and sidebar container component
